goes with grid implementation. renames some files, fixes progress bar coloring
