+++ src/components/NewLocationPage/LocationOverview/LocationOverview.stories.tsx
@@ -1,11 +1,11 @@
-import React, { Fragment } from 'react';
-import RiskOverviewFlat from './RiskOverviewFlat';
+import React from 'react';
+import LocationOverview from './LocationOverview';
 import { useLocationSummariesForFips } from 'common/hooks';
 import regions from 'common/regions';
 
 export default {
-  title: 'Location page redesign/Risk overview flat',
-  component: RiskOverviewFlat,
+  title: 'Location page redesign/Location overview',
+  component: LocationOverview,
 };
 
 export const NewYork = () => {
@@ -15,5 +15,5 @@ export const NewYork = () => {
   if (!locationSummary) {
     return null;
   }
-  return <RiskOverviewFlat region={region} locationSummary={locationSummary} />;
+  return <LocationOverview region={region} locationSummary={locationSummary} />;
 };
