--- src/components/NewLocationPage/RiskOverview/RiskOverview.stories.tsx
@@ -1,19 +1,0 @@
-import React, { Fragment } from 'react';
-import RiskOverview from './RiskOverview';
-import { useLocationSummariesForFips } from 'common/hooks';
-import regions from 'common/regions';
-
-export default {
-  title: 'Location page redesign/Risk overview',
-  component: RiskOverview,
-};
-
-export const NewYork = () => {
-  const region = regions.findByFipsCodeStrict('36');
-  const locationSummary = useLocationSummariesForFips(region.fipsCode);
-
-  if (!locationSummary) {
-    return null;
-  }
-  return <RiskOverview region={region} locationSummary={locationSummary} />;
-};
