--- src/components/NewLocationPage/VaccinationProgressBar/index.ts
@@ -1,1 +1,0 @@
-export { default as VaccinationProgressBar } from './VaccinationProgressBar';
