+++ src/components/NewLocationPage/VaccinationProgressBarBlock/ProgressBar.tsx
@@ -1,10 +1,10 @@
 import React from 'react';
 import { ParentSize } from '@vx/responsive';
 import { v4 as uuidv4 } from 'uuid';
-import { ProgressBarContainer } from './VaccinationProgressBar.style';
+import { ProgressBarContainer } from './VaccinationProgressBarBlock.style';
 import { COLOR_MAP } from 'common/colors';
 import { formatPercent } from 'common/utils';
-import { ProgressBarProps } from './VaccinationProgressBlock';
+import { ProgressBarProps } from './VaccinationProgressBarBlock';
 
 function getOffsetPercentage(decimal: number) {
   return formatPercent(decimal, 1);
@@ -42,19 +42,19 @@ const ProgressBar: React.FC<{
       <defs>
         <linearGradient id={gradientId}>
           <stop
-            offset={getOffsetPercentage(vaccinationsInitiated)}
+            offset={getOffsetPercentage(vaccinationsCompleted)}
             stopColor={initiatedFill}
           />
           <stop
+            offset={getOffsetPercentage(vaccinationsCompleted)}
+            stopColor={completedFill}
+          />
+          <stop
             offset={getOffsetPercentage(vaccinationsInitiated)}
             stopColor={completedFill}
           />
           <stop
-            offset={getOffsetPercentage(vaccinationsCompleted)}
-            stopColor={completedFill}
-          />
-          <stop
-            offset={getOffsetPercentage(vaccinationsCompleted)}
+            offset={getOffsetPercentage(vaccinationsInitiated)}
             stopColor={backgroundFill}
           />
           <stop offset="100%" stopColor={backgroundFill} />
@@ -71,7 +71,7 @@ const ProgressBar: React.FC<{
   );
 };
 
-const VaccinationProgressBarAutosize: React.FC<ProgressBarProps> = ({
+const ProgressBarAutosize: React.FC<ProgressBarProps> = ({
   vaccinationsInitiated,
   vaccinationsCompleted,
   locationName,
@@ -92,4 +92,4 @@ const VaccinationProgressBarAutosize: React.FC<ProgressBarProps> = ({
   );
 };
 
-export default VaccinationProgressBarAutosize;
+export { ProgressBarAutosize as ProgressBar };
