+++ src/components/NewLocationPage/VaccinationProgressBarBlock/index.ts
@@ -1,0 +1,3 @@
+import VaccinationProgressBarBlock from './VaccinationProgressBarBlock';
+
+export default VaccinationProgressBarBlock;
